VFB-72: Move table toolbar to separate file
