+++ src/app/admin/collectionCentresTable/CollectionCentresTableToolbar.tsx
@@ -1,0 +1,37 @@
+import {
+    GridRowModes,
+    GridRowModesModel,
+    GridRowsProp,
+    GridToolbarContainer,
+} from "@mui/x-data-grid";
+import React from "react";
+import Button from "@mui/material/Button";
+import AddIcon from "@mui/icons-material/Add";
+import { CollectionCentresTableRow } from "@/app/admin/collectionCentresTable/CollectionCentresTable";
+
+interface EditToolbarProps {
+    setRows: (newRows: (oldRows: GridRowsProp) => GridRowsProp) => void;
+    setRowModesModel: (newModel: (oldModel: GridRowModesModel) => GridRowModesModel) => void;
+    rows: CollectionCentresTableRow[];
+}
+
+export function EditToolbar(props: EditToolbarProps): React.JSX.Element {
+    const { setRows, setRowModesModel, rows } = props;
+
+    const handleClick = (): void => {
+        const id = rows.length + 1;
+        setRows((oldRows) => [...oldRows, { id, name: "", isShown: false, isNew: true }]);
+        setRowModesModel((oldModel) => ({
+            ...oldModel,
+            [id]: { mode: GridRowModes.Edit, fieldToFocus: "name" },
+        }));
+    };
+
+    return (
+        <GridToolbarContainer>
+            <Button color="primary" startIcon={<AddIcon />} onClick={handleClick}>
+                Add a new collection centre
+            </Button>
+        </GridToolbarContainer>
+    );
+}
